src/requirements.txt
django
django-debug-toolbar
django-cors-headers
django-colorfield
djangorestframework
markdown
django-filter
django-excel
xlrd==1.2.0

django-redis
gunicorn==19.9.0
django-crispy_forms
django_q
django-import-export
redis
pandas
ipython
python-snap7